demo-app/python/requirements.txt
Flask==2.0.2
flask-healthz==0.0.3
urllib3==1.26.8
python-dotenv==0.19.2
gunicorn==20.1.0
pathos==0.2.8